Pulido de funcion addProduct y optimizado de GETs

+++ src/dao/ProductManager.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const { isNumberObject } = require('util/types');
 
 class ProductManager {
 
@@ -7,23 +8,18 @@ class ProductManager {
     }
 
     async getProducts() {
-        if (fs.existsSync(this.path)) {
-
-            let products = JSON.stringify(await fs.promises.readFile(this.path, 'utf-8',(error, datoLeido)=>{
-                if(error){ console.log(`${error.message}`)
-                return}
-                console.log(datoLeido)
-            }))
+        if (fs.existsSync(this.path)) {  
+       
+            return JSON.parse(await fs.promises.readFile(this.path, 'utf-8'))
             
-            return products
         } else {
-            return ([])
+            return []           
         }
     }
 
-    async addProduct(title, description, code, price, status, stock, category, thumbnails=[]) {
+    async addProduct(title='', description='', code, price, status, stock, category='', thumbnails = []) {
 
-        let products = await this.getProducts()
+        let products = await this.getProducts(this.path)    
 
         let id = 1
 
@@ -32,17 +28,19 @@ class ProductManager {
         }
 
         let producto = {
-            id, title:(''), description:(''),code:(''), price, status, stock:(''), category:(''), thumbnails:[('')]
-        }     
-
-        if (products.find(prod=>prod.id !== producto.id)){
-
-        products.push(producto)
-        await fs.promises.writeFile(this.path, JSON.stringify(products, null, 5))
-        return producto
-    } else {
-            return console.log(`Este producto con el ID ${producto.id} ya existe`)
+            id, title, description, code, price, status, stock, category, thumbnails
         }
+
+        let productoExiste = products.find(prod => prod.id == producto.id)
+
+        if(!productoExiste) {
+
+            products.push(producto)
+            await fs.promises.writeFile(this.path, JSON.stringify(products, null, 5))
+            return producto
+        } else {
+            return `El producto ${producto.title} ya existe con el id ${producto.id}`
+        }        
     }
 }
-module.exports = { ProductManager}
+module.exports = { ProductManager }
